refactor: toStringCardHand 메서드 상속 받도록 변경
