+++ src/main/java/model/Dealer.java
@@ -7,6 +7,7 @@ import static model.Player.DELIMITER;
 
 public class Dealer extends User {
     public static final int ZERO = 0;
+
     private final Map<Result, Integer> result = new HashMap<>();
 
     public Dealer(CardHand cardHand) {
@@ -22,16 +23,6 @@ public class Dealer extends User {
 
     public Map<Result, Integer> getResult() {
         return Collections.unmodifiableMap(result);
-    }
-
-    @Override
-    public String toStringCardHand() {
-        List<String> cardNames = new ArrayList<>();
-
-        for (Card card : cardHand) {
-            cardNames.add(card.toString());
-        }
-        return String.join(DELIMITER, cardNames);
     }
 
     public void setResult(final Result result) {
