+++ src/main/java/model/Player.java
@@ -4,22 +4,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Player extends User {
-    public static final String DELIMITER = ", ";
-
     private Result result;
 
     public Player(String name, CardHand cardHand) {
         super(name, cardHand);
-    }
-
-    @Override
-    public String toStringCardHand() {
-        List<String> cardNames = new ArrayList<>();
-
-        for (Card card : cardHand) {
-            cardNames.add(card.toString());
-        }
-        return String.join(DELIMITER, cardNames);
     }
 
     public void setResult(Result result) {
@@ -29,6 +17,4 @@ public class Player extends User {
     public Result getResult() {
         return result;
     }
-
-
 }
